Scrap plans of support for API keys
There is no builtin support for AES, and I don't want to add a
dependency to an external library.

Also the API key functionality in AniDB isn't really an ordinary API
key. The user password has to be in the configuration file even if an
API key is specified.

The main benefit of using an API key is that it enables encryption which
could allow us not to send the password in clear text.